Change default depth to num_segs instead of log_2(num_segs)
log_2(num_segs) made sense for the kd_tree, because each node of the tree
is always split exactly in half, resulting in a balanced binary tree.
With the optimized tree algorithm, splits can happen anywhere, there can
be 1 segment on the left and 50 on the right, and then later down the
tree the 50 get split up more while the 1 does not.  In the worst case,
one node gets split off from the rest at each cut, meaning in order to
populate num_segs leaves you need num_segs cuts.  This could result in
a longer optimzation time, but the optimization phase is usually very
fast compared to the actual parallel-DBSCAN phase that follows.
It should only matter for small datasets, and the user always has the
option of setting a lower max_depth than the default if optimization
is taking too much time.